chore: Renamed menu bar and related components
