+++ packages/frontend/src/components/MenuBar/index.tsx
@@ -1,4 +1,4 @@
-export * from './MenuBar.tsx';
+export * from './NavigationMenuBar.tsx';
 export * from '../Backdrop/Backdrop.tsx';
-export * from './DropDownMenu.tsx';
-export * from './DropDownSubMenuInbox.tsx';
+export * from './NavigationDropdownMenu.tsx';
+export * from './NavigationDropdownSubMenuInbox.tsx';
